🔨 Refactor(#83): 스타일 파일 위치 이동
