+++ src/features/store/ui/ItemContainer.tsx
@@ -1,4 +1,4 @@
-import * as S from '../styles';
+import * as S from './styles';
 import { useState } from 'react';
 import type Item from '@type/Item';
 import StoreItem from './StoreItem';
